Fixed rotation of fired hook.
This bug is fixed by addressing 2 issues.

The first is making sure the rotation of the hook is maintained when detaching the hook actor from the handle actor. This is done by changing the `Rotation Rule` parameter in the `DetachFromActor` node in the `Update Position` subgraph in the `BP_Hook2` blueprint from `Keep Relative` to `Keep World`.

The second is modifying the unit directional vector of the drop point and the target location. The hook model was created with the wrong axis as the forward direction, this is simply fixed programmatically. Cross product the directional vector with down to rotate it correctly.