Hook can only deploy when laser is ready.
When equipping the laser pointer in Deep Rock Galactic, there is a delay until the actual laser shows up and scans the terrain. Firing the hook before the laser was visible would have unexpected behavior, such as firing away from the cursor to the last known laser hit. This commit fixes this issue by not letting the player fire the hook until the laser pointer is completely equipped.